backend/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
bandit==1.9.2
boto3==1.42.19
botocore==1.42.19
cffi==2.0.0
decorator==5.2.1
dj-database-url==2.1.0
Django==5.0.14
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.1
django-ratelimit==4.1.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
gunicorn==21.2.0
inflection==0.5.1
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rich==14.2.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
stack-data==0.6.3
stevedore==5.6.0
traitlets==5.14.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
wcwidth==0.2.14
whitenoise==6.6.0
